Rating bad fix

+++ recycleIT-backend/src/main/java/com/recycleit/recycleitbackend/service/RatingService.java
@@ -6,6 +6,6 @@ import java.math.BigDecimal;
 
 public interface RatingService {
     void setRating(RatingDto rating);
-    void changeRating(Long serviceId,Long userId, int mark);
+    void changeRating(Long serviceId,Long userId, BigDecimal mark);
     BigDecimal getFacilityRating(Long serviceId);
 }
